requirements.txt
streamlit
pandas
openpyxl
Pillow
arabic_reshaper
python-bidi